requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-filter==22.1
django-selenium-login==2.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.4
h11==0.13.0
idna==3.3
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
pyparsing==3.0.9
PySocks==1.7.1
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
zipp==3.10.0

